requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
scrapy==2.11.0
requests==2.31.0
